Extract get data function from createInputsFormFromFilter. Each utils method needs to do just one job.
