+++ src/utils/createInputsFormFromFilter.js
@@ -4,82 +4,19 @@ import DataRouterService                                 from 'services/data';
 import { toRawType }                                     from 'utils/toRawType';
 import { getUniqueDomId }                                from 'utils/getUniqueDomId';
 import { createFilterFormInputs }                        from 'utils/createFilterFormInputs';
-import { createFieldsDependenciesAutocompleteParameter } from 'utils/createFieldsDependenciesAutocompleteParameter';
-import { sortAlphabeticallyArray }                       from 'utils/sortAlphabeticallyArray';
-import { sortNumericArray }                              from 'utils/sortNumericArray';
 import { getUniqueValuesFromField }                      from "utils/getUniqueValuesFromField";
+import { getDataForSearchInput }                          from "./getDataForSearchInput";
 
 /**
  * Create the right search structure for search form
  * 
  * @param { Object } opts
  * @param { Object } opts.state
- * @param { Object } opts.fromField
+ * @param { Object } opts.fromField {field, value, output}
  * 
- * @returns { Promise<void> } form input
+ * @returns { Promise<Array|void> } form input
  */
-export async function createInputsFormFromFilter({ state, fromField }) {
-
-  // Get unique values from field
-  if (fromField) {
-    let { field, value, output } = fromField;
-
-    let data = [];
-    try {
-
-      // get current field dependence
-      let dep = state.input.dependance[fromField.field];
-      if (dep && (state.input.cached_deps[dep] && SEARCH_ALLVALUE !== state.input.cached_deps[dep]._currentValue)) {
-        dep = { [dep]: state.input.cached_deps[dep]._currentValue }; // dependance as value
-      } else if(dep) {
-        dep = { [dep]: undefined }; // undefined = so it no add on list o field dependance
-      }
-
-      let autoFieldDependecies;
-
-      if (dep) {
-        const [field, value] = Object.entries(dep)[0];
-        autoFieldDependecies = createFieldsDependenciesAutocompleteParameter({
-          field,
-          value,
-          filter: state.filter,
-          inputdependance: state.input.dependance,
-          cachedependencies: state.input.cached_deps,
-        })
-      }
-
-      const layers = (1 === state.search_layers.length ? [state.search_layers[0]] : state.search_layers);
-      const response = Array.from(
-        (
-          await Promise
-          .allSettled(layers.map(l => l.getFilterData({
-            field:      autoFieldDependecies,
-            suggest:    value !== undefined ? `${field}|${value}` : undefined,
-            unique:     field,
-            ordering:   field,
-            fformatter: undefined,
-          })))
-        )
-          .filter(d => 'fulfilled' === d.status)
-          .reduce((acc, { value = [] }) => new Set([...acc, ...value]), [])
-      )
-
-      // check if is not an empty array
-      switch (response.length && typeof response[0]) {
-        case 'string': data = sortAlphabeticallyArray(response); break;
-        case 'number': data = sortNumericArray(response); break;
-        default:       data = response;
-      }
-
-      if ('autocomplete' === output) {
-        data = data.map(d => ({ id: d, text: d }));
-      }
-    } catch(e) {
-      console.warn(e);
-    }
-
-    return data;
-  }
+export async function createInputsFormFromFilter({ state }) {
 
   const inputdependance   = state.input.dependance;
   const inputdependencies = state.input.dependencies;
